fix: draft button removed from feedback + renames files to add clarity BREAKING CHANGE: form renamed and refactored, creates new ReportForm componentthat extends from Form component
